+++ config-ui/Common/src/main/java/de/markusbordihn/easynpc/configui/network/message/server/ResetExecutionLimitMessage.java
@@ -20,16 +20,17 @@
 package de.markusbordihn.easynpc.configui.network.message.server;
 
 import de.markusbordihn.easynpc.configui.Constants;
+import de.markusbordihn.easynpc.data.execution.ExecutionId;
+import de.markusbordihn.easynpc.data.execution.ExecutionType;
 import de.markusbordihn.easynpc.data.saveddata.ActionExecutionTracker;
 import de.markusbordihn.easynpc.network.message.NetworkMessageRecord;
-import java.util.UUID;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public record ResetExecutionLimitMessage(UUID executionId, boolean allPlayers)
+public record ResetExecutionLimitMessage(ExecutionId executionId, boolean allPlayers)
     implements NetworkMessageRecord {
 
   public static final ResourceLocation MESSAGE_ID =
@@ -37,12 +38,15 @@ public record ResetExecutionLimitMessage(UUID executionId, boolean allPlayers)
   private static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
 
   public static ResetExecutionLimitMessage create(final FriendlyByteBuf buffer) {
-    return new ResetExecutionLimitMessage(buffer.readUUID(), buffer.readBoolean());
+    return new ResetExecutionLimitMessage(
+        new ExecutionId(buffer.readEnum(ExecutionType.class), buffer.readUUID()),
+        buffer.readBoolean());
   }
 
   @Override
   public void write(final FriendlyByteBuf buffer) {
-    buffer.writeUUID(this.executionId);
+    buffer.writeEnum(this.executionId.type());
+    buffer.writeUUID(this.executionId.value());
     buffer.writeBoolean(this.allPlayers);
   }