+++ core/Forge/src/main/java/de/markusbordihn/easynpc/gametest/ConditionEvaluationTest.java
@@ -60,4 +60,10 @@ public class ConditionEvaluationTest {
     ConditionEvaluationTestHelper.assertConditionalDialogOpenRespectsConditions(helper);
     helper.succeed();
   }
+
+  @GameTest(template = "gametest.3x3x3")
+  public void testDefaultDialogExecutionLimitEnforced(GameTestHelper helper) {
+    ConditionEvaluationTestHelper.assertDefaultDialogExecutionLimitEnforced(helper, humanoid());
+    helper.succeed();
+  }
 }